% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bdcaltenderheat}
\alias{bdcaltenderheat}
\title{bdcalenderheat - Calender Heat map of biodiversity data}
\usage{
bdcaltenderheat(indf = NA, title = NA)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{title}{title field name for the plot}
}
\description{
bdcalenderheat - Calender Heat map of biodiversity data
}
\examples{
\dontrun{
bdcalenderheat(inat)
}
}

