\name{bdsummary}
\alias{bdsummary}
\title{bdsummary - description}
\usage{
  bdsummary(indf)
}
\arguments{
  \item{indf}{input data frame containing biodiversity data
  set}
}
\description{
  bdsummary - description
}
\examples{
\dontrun{
require(rinat)
inat<-get_obs_project("reptileindia")
bdsummary(inat)
}
}

