% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdvis.R
\docType{package}
\name{bdvis}
\alias{bdvis}
\alias{bdvis-package}
\title{bdvis: Biodiversity Data Visualizations}
\description{
Biodiversity data visualizations using R would be helpful to understand
completeness of biodiversity inventory, extent of geographical, taxonomic
and temporal coverage, gaps and biases in data.
}
\section{Data preparation}{

\itemize{
\item \link{bd_summary}
\item \link{fix_str} Deprecated
\item \link{format_bdvis}
\item \link{get_cell_id}
\item \link{get_taxo}
}
}

\section{Spatial visualizations}{

\itemize{
\item \link{bd_webmap}
\item \link{map_grid}
}
}

\section{Temporal Visualizations}{

\itemize{
\item \link{bd_calendarheat}
\item \link{chronohorogram}
\item \link{tempolar}
}
}

\section{Taxonomic Visualizations}{

\itemize{
\item \link{taxo_tree}
}
}

\section{Miscellaneous functions}{

\itemize{
\item \link{bd_complete}
\item \link{distri_graph}
}
}

\section{Citation}{

\itemize{
\item Barve, V., & Otegui, J. (2016). bdvis: Biodiversity data
visualizations (R package V 0.2). Retrieved from
https://cran.r-project.org/web/packages/bdvis/index.html
}
}

