% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdwebmap.R
\name{bdwebmap}
\alias{bdwebmap}
\title{(Deprecated) Interactive web page based map of records}
\usage{
bdwebmap()
}
\value{
No return value. NULL
}
\description{
(Deprecated) Interactive web page based map of records
}
