% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasubset.R
\name{datasubset}
\alias{datasubset}
\title{Subset data for bdvis}
\usage{
datasubset(indf = NA, sql = "", minyear = -1, maxyear = -1,
  scname = NA)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{sql}{sql string to run query on indf}

\item{minyear}{Minimum year value}

\item{maxyear}{Maximum year value}

\item{scname}{Scientific_name to select}
}
\description{
Subset data for bdvis
}
\examples{
\dontrun{
inat1=datasubset(inat,minyear=2000, maxyear=2014)
}
}

