% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixstr.R
\name{fix_str}
\alias{fix_str}
\title{Change structure of the data frame according to the package's needs}
\usage{
fix_str(indf, Latitude = NA, Longitude = NA, date_collected = NA,
  datefmt = NA, sci_name = NA)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{Latitude}{name of the Latitude field in original data frame}

\item{Longitude}{name of the Longitude field in original data frame}

\item{date_collected}{name of the Date Collected field in original data frame}

\item{datefmt}{format string for the original date field \link[base]{strptime}}

\item{sci_name}{name of the Scientific Name field in original data frame}
}
\description{
This function is slated to deprecate in next version. Please use function
\link{format_bdvis} instead.
}
\details{
Modify the name of certain fields in the provided data.frame to meet the
requirements of the package, to make sure functions are executed properly.
}
\examples{
\dontrun{
inat <- fix_str(inat, date_collected = "Date.collected", datefmt = "\%Y-\%m-\%d \%H:\%M:\%S")
}
}
\seealso{
Other Data preparation functions: \code{\link{bd_summary}},
  \code{\link{format_bdvis}}, \code{\link{get_cell_id}},
  \code{\link{get_taxo}}
}
\concept{Data preparation functions}
