% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixstr.R
\name{fixstr}
\alias{fixstr}
\title{Change structure of the data frame according to the package's needs}
\usage{
fixstr(indf, Latitude = NA, Longitude = NA, DateCollected = NA,
  datefmt = NA, SciName = NA)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{Latitude}{name of the Latitude field in original data frame}

\item{Longitude}{name of the Longitude field in original data frame}

\item{DateCollected}{name of the Date Collected field in original data frame}

\item{datefmt}{format string for the original date field \link[base]{strptime}}

\item{SciName}{name of the Scientific Name field in original data frame}
}
\description{
This function is slated to deprecate in next version. Please use function
\link{format_bdvis} instead.
}
\details{
Modify the name of certain fields in the provided data.frame to meet the
requirements of the package, to make sure functions are executed properly.
}
\examples{
\dontrun{
inat = fixstr(inat, DateCollected = "Date.collected", datefmt = "\%Y-\%m-\%d \%H:\%M:\%S")
}
}
\seealso{
Other Data preparation functions: \code{\link{bdsummary}},
  \code{\link{format_bdvis}}, \code{\link{getcellid}},
  \code{\link{gettaxo}}
}
