% Generated by roxygen2 (4.0.0): do not edit by hand
\name{format_bdvis}
\alias{format_bdvis}
\title{format_bdvis - Convert data to bdvis native format}
\usage{
format_bdvis(indf = NA, gettaxo = F)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{gettaxo}{Run gettaxo function to get higher taxonomy if true}
}
\description{
format_bdvis - Convert data to bdvis native format
}
\examples{
\dontrun{
inat=format_bdvis(inat)
}
}

