% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcellid.R
\name{get_cell_id}
\alias{get_cell_id}
\title{Assign GBIF style degree cell ids}
\usage{
get_cell_id(indf)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}
}
\description{
Calculate and assign a GBIF-style degree cell id and centi-degree (0.1
degrees, dividing a 1 degree cell into 100 centi-degree cells) cell id to each
record. This is a necessary previous step for some functions like
\code{\link{map_grid}}
}
\examples{
\dontrun{
get_cell_id(inat)
}
}
\seealso{
Other Data preparation functions: \code{\link{bd_summary}},
  \code{\link{fix_str}}, \code{\link{format_bdvis}},
  \code{\link{get_taxo}}
}
\concept{Data preparation functions}
