% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapgrid.R
\name{mapgrid}
\alias{mapgrid}
\title{mapgrid - Maps the data points on the map in grid format}
\usage{
mapgrid(indf = NA, ptype = "records", bbox = NA, title = "",
  mapdatabase = "world", region = ".", customize = NULL)
}
\arguments{
\item{indf}{- input data frame containing biodiversity data set}

\item{ptype}{- type of map on the grid valid values are presence, records, species}

\item{bbox}{- Bounding box for the map in format c(xmin,xmax,ymin,ymax)}

\item{title}{- title for the map}

\item{mapdatabase}{- database to be used default world}

\item{region}{- specify region(s) to map i.e. countries default . for whole world}

\item{customize}{- customization string}
}
\description{
mapgrid - Maps the data points on the map in grid format
}
\examples{
\dontrun{
mapgrid(inat,ptype="records")
}
}

