% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gui.R
\name{run_gui}
\alias{run_gui}
\title{Run a Graphical User Interface (GUI) for BDVis}
\usage{
run_gui()
}
\description{
This function starts a browser based application that exposes the major functions of the package.
}
\examples{
run_gui()

}
\author{
Boyan Angelov
}
