\name{summary}
\alias{summary}
\title{summary - description}
\usage{
  summary(indf)
}
\arguments{
  \item{indf}{input data frame containing biodiversity data
  set}
}
\description{
  summary - description
}
\examples{
\dontrun{
require(rinat)
inat<-get_obs_project("reptileindia")
summary(inat)
}
}

