\name{tempolar}
\alias{tempolar}
\title{tempolar - Polar plot of temporal data}
\usage{
  tempolar(indf = NA, timescale = NA, title = NA,
    color = NA, plottype = NA)
}
\arguments{
  \item{indf}{input data frame containing biodiversity data
  set}

  \item{timescale}{Temporal scale of the graph d - daily, w
  - weekly m - monthly}

  \item{title}{Title for the graph}

  \item{color}{color of the graph plot}

  \item{plottype}{plot types of r - lines, p - polygon and
  s - symbols}
}
\description{
  tempolar - Polar plot of temporal data
}
\examples{
\dontrun{
tempolar(inat)
}
}

