% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdcalendarheat.R
\name{bdcalendarheat}
\alias{bdcalendarheat}
\title{Calendar heat map of biodiversity data}
\usage{
bdcalendarheat(indf = NA, title = NA)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{title}{title custome title for the plot}
}
\description{
Produces a heat map {\url{https://en.wikipedia.org/wiki/Heat_map}}
representing the distribution of records in time.
}
\details{
The calendar heat map is a matrix-like plot where each cell represents a 
unique date, and the color the cell is painted with shows the amount of 
records that have that particular date. Rows are weekdays and columns are 
week numbers, each year having its own "panel".
}
\examples{
\dontrun{
bdcalendarheat(inat)
}
}
\seealso{
Other Temporal visualizations: \code{\link{chronohorogram}},
  \code{\link{tempolar}}
}

