% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdcalendarheat.R
\name{bdcalendarheat}
\alias{bdcalendarheat}
\title{bdcalendarheat - Calendar Heat map of biodiversity data}
\usage{
bdcalendarheat(indf = NA, title = NA)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{title}{title field name for the plot}
}
\description{
bdcalendarheat - Calendar Heat map of biodiversity data
}
\examples{
\dontrun{
bdcalendarheat(inat)
}
}

