\name{bdcomplete}
\alias{bdcomplete}
\title{bdcomplete - Computes completeness values for each cell currently returns Chao2}
\usage{
  bdcomplete(indf, recs = 50)
}
\arguments{
  \item{indf}{input data frame containing biodiversity data
  set}

  \item{recs}{Minimum number of records per grid cell}
}
\description{
  bdcomplete - Computes completeness values for each cell
  currently returns Chao2
}
\examples{
\dontrun{
bdcomplete(inat)
}
}

