% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdcomplete.R
\name{bdcomplete}
\alias{bdcomplete}
\title{bdcomplete - Computes completeness values for each cell currently returns Chao2}
\usage{
bdcomplete(indf, recs = 50)
}
\arguments{
\item{indf}{- Input data frame containing biodiversity data set}

\item{recs}{- Minimum number of records per grid cell 
(Default is 50, if the number is too low, might give error)}
}
\value{
data.frame with the columns
\itemize{
 \item{"Cell_id"}{id of the cell}
 \item{"Sobs"}{Number of Observed species}
 \item{"Sest"}{Estimated number of species}
 \item{"c"}{Completeness ratio the cell}
 Plots a graph of Number of species vs completeness
}
}
\description{
bdcomplete - Computes completeness values for each cell currently returns Chao2
}
\examples{
\dontrun{
bdcomplete(inat)
}
}

