% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bdsummary}
\alias{bdsummary}
\title{bdsummary - Provides summary of data like number of records, Families,
          Genus, Species, Bounding box of locations, date range and
          spatial coverage.}
\usage{
bdsummary(indf)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}
}
\description{
bdsummary - Provides summary of data like number of records, Families,
          Genus, Species, Bounding box of locations, date range and
          spatial coverage.
}
\examples{
\dontrun{
require(rinat)
inat<-get_inat_obs_project("reptileindia")
bdsummary(inat)
}
}

