% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bdsummary}
\alias{bdsummary}
\title{bdsummary - description}
\usage{
bdsummary(indf)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}
}
\description{
bdsummary - description
}
\examples{
\dontrun{
require(rinat)
inat<-get_inat_obs_project("reptileindia")
bdsummary(inat)
}
}

