% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdvis.R
\docType{package}
\name{bdvis}
\alias{bdvis}
\alias{bdvis-package}
\title{bdvis: Biodiversity Data Visualizations}
\description{
Biodiversity data visualizations using R would be helpful to understand 
completeness of biodiversity inventory, extent of geographical, taxonomic 
and temporal coverage, gaps and biases in data.
}
\section{Data preparation}{

\itemize{  
\item \link{bdsummary}
\item \link{fixstr}
\item \link{format_bdvis}
\item \link{getcellid}
\item \link{gettaxo}
}
}

\section{Spatial visualizations}{

\itemize{ 
\item \link{bdwebmap}
\item \link{mapgrid}
}
}

\section{Temporal Visualizations}{

\itemize{  
\item \link{bdcalendarheat}
\item \link{chronohorogram}
\item \link{tempolar}
}
}

\section{Taxonomic Visualizations}{

\itemize{  
\item \link{taxotree}
}
}

\section{Miscellaneous functions}{

\itemize{  
\item \link{bdcomplete}
\item \link{distrigraph}
}
}

\section{Citation}{

\itemize{  
\item Barve, V., & Otegui, J. (2016). bdvis: Biodiversity data 
visualizations (R package V 0.1). Retrieved from 
https://cran.r-project.org/web/packages/bdvis/index.html
}
}

