% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdwebmap.R
\name{bdwebmap}
\alias{bdwebmap}
\title{Interactive web page based map of records}
\usage{
bdwebmap(indf)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}
}
\description{
Generates a interactive web map and opens it in a new browser window. To make 
this map meaningful the number of species restricted to less than 30 and number 
of map points to less than 1000.
}
\examples{
\dontrun{
 bdwebmap(inat)
}
}
\seealso{
Other Spatial visualizations: \code{\link{mapgrid}}
}

