% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdwebmap.R
\name{bdwebmap}
\alias{bdwebmap}
\title{bdwebmap - Interactive web page based map of records}
\usage{
bdwebmap(indf)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}
}
\description{
bdwebmap - Interactive web page based map of records
}
\examples{
\dontrun{
 bdwebmap(inat)
}
}

