% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronohorogram.R
\name{chronohorogram}
\alias{chronohorogram}
\title{Draws a chronohorogram of records}
\usage{
chronohorogram(indf = NA, title = NA, startyear = 0, endyear = 0,
  colors = c("red", "blue"))
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{title}{title of the plot}

\item{startyear}{starting year for the plot}

\item{endyear}{end year for the graph}

\item{colors}{colors to build color ramp}
}
\description{
Draws a chronohorogram of records
}
\examples{
\dontrun{
chronohorogram(inat)
}
}
\references{
Arino, A. H., & Otegui, J. (2008). Sampling biodiversity sampling. In Proceedings of TDWG (pp. 77-78). Retrieved from http://www.tdwg.org/fileadmin/2008conference/documents/Proceedings2008.pdf#page=77
}

