% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrigraph.R
\name{distrigraph}
\alias{distrigraph}
\title{Distribution graphs}
\usage{
distrigraph(indf, ptype = NA, ...)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{ptype}{type of graph i.e. cell, species, efforts}

\item{...}{any additional parameters for plot. Standard papramaterrs like graph type, color etc.}
}
\description{
Graphs dispalying distribution of biodiversity data
}
\examples{
\dontrun{
 distrigraph(inat2,type="cell")
 distrigraph(inat2,type="species")
 distrigraph(inat2,type="efforts",col="red")
}
}

