% Generated by roxygen2 (4.0.0): do not edit by hand
\name{distrigraph}
\alias{distrigraph}
\title{distrigraph - Distribution graphs}
\usage{
distrigraph(indf, type = NA)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{type}{type of graph i.e. cell, species, efforts}
}
\description{
distrigraph - Distribution graphs
}
\examples{
\dontrun{
 distrigraph(inat2,type="cell")
 distrigraph(inat2,type="species")
 distrigraph(inat2,type="efforts")
}
}

