% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrigraph.R
\name{distrigraph}
\alias{distrigraph}
\title{distrigraph - Distribution graphs}
\usage{
distrigraph(indf, type = NA, ...)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{type}{type of graph i.e. cell, species, efforts}

\item{...}{any additional parameters for plot}
}
\description{
distrigraph - Distribution graphs
}
\examples{
\dontrun{
 distrigraph(inat2,type="cell")
 distrigraph(inat2,type="species")
 distrigraph(inat2,type="efforts",col="red")
}
}

