% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_bdvis.R
\name{format_bdvis}
\alias{format_bdvis}
\title{Convert data to bdvis native format}
\usage{
format_bdvis(indf = NA, gettaxo = F)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{gettaxo}{Run gettaxo function to get higher taxonomy if true}
}
\description{
Convert data to bdvis native format so that other bdvis functions understand the data
}
\examples{
\dontrun{
inat=format_bdvis(inat)
}
}

