% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcellid.R
\name{getcellid}
\alias{getcellid}
\title{Assign GBIF style degree cell ids}
\usage{
getcellid(indf)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}
}
\description{
Calculate and assign a GBIF-style degree cell id and centi-degree (0.1 
degrees, dividing a 1 degree cell into 100 centi-degree cells) cell id to each
record. This is a necessary previous step for some functions like
\code{\link{mapgrid}}
}
\examples{
\dontrun{
getcellid(inat)
}
}
\seealso{
Other Data preparation functions: \code{\link{bdsummary}},
  \code{\link{fixstr}}, \code{\link{gettaxo}}
}

