% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcellid.R
\name{getcellid}
\alias{getcellid}
\title{Assigns GBIF style degree cell ids}
\usage{
getcellid(indf)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}
}
\description{
Assigns GBIF style degree cell ids and Centi degree (0.1) cell ids for each record which are used in functions like \link[bdvis]{mapgrid}
}
\examples{
\dontrun{
getcellid(inat)
}
}

