% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcellid.R
\name{getcellid}
\alias{getcellid}
\title{getcellid - Assigns GBIF style degree Cell ids and Centi degree (0.1) cell ids 
 for each record.}
\usage{
getcellid(indf)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}
}
\description{
getcellid - Assigns GBIF style degree Cell ids and Centi degree (0.1) cell ids 
 for each record.
}
\examples{
\dontrun{
getcellid(inat)
}
}

