% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettaxo.R
\name{gettaxo}
\alias{gettaxo}
\title{gettaxo - Get higher taxonomy fields data}
\usage{
gettaxo(indf, genus = FALSE)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{genus}{If true use only genus level data to get taxanomy}
}
\value{
indf with added / updated columns
\itemize{
 \item{"kingdom"}{Kingdom of the Scientific name}
 \item{"phylum"}{Phylum of the Scientific name}
 \item{"order"}{Order of the Scientific name}
 \item{"family"}{Family of the Scientific name}
 \item{"genus"}{Genus of the Scientific name}
}
and also saves a local copy of taxanomy downloaded for future use in taxo.bd sqlite file,
}
\description{
gettaxo - Get higher taxonomy fields data
}
\examples{
\dontrun{
inat=gettaxo(inat)
}
}

