% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapgrid.R
\name{mapgrid}
\alias{mapgrid}
\title{Maps the data points on the map in grid format}
\usage{
mapgrid(indf = NA, ptype = "records", title = "", bbox = NA,
  legscale = 0, mapdatabase = "world", region = ".", customize = NULL)
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{ptype}{plot type of map on the grid valid values are presence, records, species. Presence will generate presence maps, species will display number of species in each map pixel and records will display number of records in each map pixel.}

\item{title}{title for the map}

\item{bbox}{bounding box for the map in format c(xmin,xmax,ymin,ymax)}

\item{legscale}{Set legend scale to a higher value than the max value in the data}

\item{mapdatabase}{database to be used default world}

\item{region}{specify region(s) to map i.e. countries default '.' for whole world map}

\item{customize}{additional customization string to customize the map output using ggplo2 parameters}
}
\description{
Maps the data points aggrigated on the map in grid format showing presence 
  or number of records in each gid cell by color gradiant.
}
\examples{
\dontrun{
mapgrid(inat,ptype="records")
}
}

