\name{taxotree}
\alias{taxotree}
\title{taxotree - Draws a treemap based on Taxonomic hierarchy of records}
\usage{
  taxotree(indf)
}
\arguments{
  \item{indf}{input data frame containing biodiversity data
  set}
}
\description{
  taxotree - Draws a treemap based on Taxonomic hierarchy
  of records
}
\examples{
\dontrun{
taxotree(inat)
}
}

