\name{taxotree}
\alias{taxotree}
\title{taxotree - Draws a treemap based on Taxonomical hirarchi of records}
\usage{
  taxotree(indf)
}
\arguments{
  \item{indf}{input data frame containing biodiversity data
  set}
}
\description{
  taxotree - Draws a treemap based on Taxonomical hirarchi
  of records
}
\examples{
\dontrun{
treemap(inat)
}
}

