% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxotree.R
\name{taxotree}
\alias{taxotree}
\title{taxotree - Draws a treemap based on Taxonomic hierarchy of records}
\usage{
taxotree(indf, n = 30, title = NA, legend = NA, sum1 = "Family",
  sum2 = "Genus")
}
\arguments{
\item{indf}{input data frame containing biodiversity data set}

\item{n}{max number of rectangles to be plotted in the treemap}

\item{title}{title for the tree}

\item{legend}{legend title}

\item{sum1}{Taxanomic fieldname to summarize data level 1}

\item{sum2}{Taxanomic fieldname to summarize data level 2}
}
\description{
taxotree - Draws a treemap based on Taxonomic hierarchy of records
}
\examples{
\dontrun{
taxotree(inat)
}
}
\references{
Otegui, J., Arino, A. H., Encinas, M. A., & Pando, F. (2013). Assessing the Primary Data Hosted by the Spanish Node of the Global Biodiversity Information Facility (GBIF). PLoS ONE, 8(1), e55144. doi:10.1371/journal.pone.0055144
}

